Extract enums to a separate unit
